Remove unecessary double Phoenix namespace
and what was called Phoenix is now Socket
