--- Sources/Phoenix/Phoenix/Reply.swift
@@ -1,16 +1,0 @@
-import Foundation
-
-extension Phoenix {
-    public struct Reply {
-        internal let joinRef: Ref
-        internal let ref: Ref
-        
-        let message: Message
-        
-        let status: String
-        let response: Dictionary<String, Any>
-        
-        public var isOk: Bool { return status == "ok" }
-        public var isNotOk: Bool { return isOk == false }
-    }
-}
